requirements.txt
folium==0.14.0
haversine==2.8.0
numpy==1.25.0
pandas==2.0.3
patsy==0.5.3
Pillow==9.5.0
plotly==5.15.0
plotly-express==0.4.1
streamlit==1.23.1
streamlit-folium==0.12.0
